Modify MonitorActivity

+++ app/src/main/java/team5/cs560/kaist/cs560team5/ListenerService.java
@@ -278,7 +278,7 @@ public class ListenerService extends Service  {
                 editor.putString("monName", (String) tuples[0]);
                 editor.putFloat("monLa", ((Double)tuples[1]).floatValue());
                 editor.putFloat("monLo", ((Double)tuples[2]).floatValue());
-                editor.putLong("monHr", (Integer) tuples[3]);
+                editor.putLong("monHr", (Long) tuples[3]);
             }
             editor.commit();
         }
@@ -289,7 +289,7 @@ public class ListenerService extends Service  {
             table.reset();
             while (table.hasNext()) {
                 tuples = table.getTuple();
-                hr = (Integer)tuples[3];
+                hr = (Long)tuples[3];
             }
             return hr;
         }
@@ -309,6 +309,7 @@ public class ListenerService extends Service  {
         }
 
         private void startDistNoti(){
+
             NotificationManager nm2 = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);   // call notification manager
             Notification notification2 = new Notification(R.mipmap.distance, "Warning! far from you!", System.currentTimeMillis());   // icon, tickerText, when
             notification2.flags = Notification.FLAG_AUTO_CANCEL;
